# api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentic AI Agricultural Crop Planning & Precision Farming System",
    "summary": "Agentic multi-agent crop planning and precision farming API.",
    "description": "\nMulti-agent decision support for crop planning and precision farming.\n\nThe workflow is orchestrated with **LangGraph** over twelve specialised agents\n(farm profile, soil & nutrient, weather & climate, agricultural knowledge RAG,\ncrop suitability, ML forecasting, irrigation, crop-risk review, activity\nplanning and advisory). Every agent emits structured evidence, and every\nphysical action (irrigation in particular) requires **human approval** before it\ncan leave the `planned` state.\n",
    "contact": {
      "name": "Precision Farming Platform"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/sensors/readings": {
      "get": {
        "tags": [
          "sensors"
        ],
        "summary": "List sensor readings",
        "operationId": "list_readings_api_v1_sensors_readings_get",
        "parameters": [
          {
            "name": "field_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Field Id"
            }
          },
          {
            "name": "sensor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sensor Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SensorReadingOut"
                  },
                  "title": "Response List Readings Api V1 Sensors Readings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sensors"
        ],
        "summary": "Ingest a device reading",
        "operationId": "create_reading_api_v1_sensors_readings_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorReadingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorReadingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sensors/fields/{field_id}/latest": {
      "get": {
        "tags": [
          "sensors"
        ],
        "summary": "Latest reading",
        "operationId": "latest_reading_api_v1_sensors_fields__field_id__latest_get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          },
          {
            "name": "sensor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sensor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SensorReadingOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Latest Reading Api V1 Sensors Fields  Field Id  Latest Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sensors/fields/{field_id}/simulate": {
      "post": {
        "tags": [
          "sensors"
        ],
        "summary": "Generate simulated telemetry (every row is flagged as simulated)",
        "operationId": "simulate_api_v1_sensors_fields__field_id__simulate_post",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorSimulationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorSimulationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sensors/fields/{field_id}/trend": {
      "get": {
        "tags": [
          "sensors"
        ],
        "summary": "Telemetry trend and statistics",
        "operationId": "trend_api_v1_sensors_fields__field_id__trend_get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2160,
              "minimum": 1,
              "default": 168,
              "title": "Hours"
            }
          },
          {
            "name": "sensor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sensor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorTrend"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sensors/fields/{field_id}/quality": {
      "get": {
        "tags": [
          "sensors"
        ],
        "summary": "Data-quality audit for the latest telemetry",
        "operationId": "quality_api_v1_sensors_fields__field_id__quality_get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Quality Api V1 Sensors Fields  Field Id  Quality Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SensorReadingCreate": {
        "properties": {
          "field_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Id"
          },
          "sensor_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Sensor Id",
            "default": "SM-01"
          },
          "recorded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded At"
          },
          "soil_moisture_percent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Moisture Percent"
          },
          "soil_temperature_c": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 80.0,
                "minimum": -20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Temperature C"
          },
          "air_humidity_percent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Air Humidity Percent"
          },
          "air_temperature_c": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 70.0,
                "minimum": -30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Air Temperature C"
          },
          "battery_percent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Percent"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": false
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "SensorReadingCreate"
      },
      "SensorReadingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "field_id": {
            "type": "integer",
            "title": "Field Id"
          },
          "sensor_id": {
            "type": "string",
            "title": "Sensor Id"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "soil_moisture_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Moisture Percent"
          },
          "soil_temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Temperature C"
          },
          "air_humidity_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Air Humidity Percent"
          },
          "air_temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Air Temperature C"
          },
          "battery_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Percent"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated"
          },
          "quality_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quality Flags"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "field_id",
          "sensor_id",
          "recorded_at",
          "soil_moisture_percent",
          "soil_temperature_c",
          "air_humidity_percent",
          "air_temperature_c",
          "battery_percent",
          "is_simulated",
          "quality_flags",
          "notes"
        ],
        "title": "SensorReadingOut"
      },
      "SensorSimulationRequest": {
        "properties": {
          "hours": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 6.0,
            "title": "Hours",
            "description": "Window length to simulate",
            "default": 72
          },
          "interval_hours": {
            "type": "number",
            "maximum": 24.0,
            "exclusiveMinimum": 0.0,
            "title": "Interval Hours",
            "default": 3.0
          },
          "sensor_id": {
            "type": "string",
            "maxLength": 64,
            "title": "Sensor Id",
            "default": "SM-01"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start At"
          },
          "initial_soil_moisture_percent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Soil Moisture Percent"
          },
          "rainfall_events": {
            "type": "boolean",
            "title": "Rainfall Events",
            "description": "Include a rainfall event so downstream rain-adaptation logic is exercised",
            "default": true
          },
          "inject_fault": {
            "type": "boolean",
            "title": "Inject Fault",
            "description": "Inject one sensor fault for alert testing",
            "default": false
          },
          "replace_existing": {
            "type": "boolean",
            "title": "Replace Existing",
            "description": "Delete this sensor's existing SIMULATED readings inside the window before generating. Required when simulated telemetry already overlaps the window - otherwise the request is refused, because two overlapping series produce duplicated timestamps and a meaningless trend.",
            "default": false
          }
        },
        "type": "object",
        "title": "SensorSimulationRequest"
      },
      "SensorSimulationResponse": {
        "properties": {
          "field_id": {
            "type": "integer",
            "title": "Field Id"
          },
          "sensor_id": {
            "type": "string",
            "title": "Sensor Id"
          },
          "readings_created": {
            "type": "integer",
            "title": "Readings Created"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "readings": {
            "items": {
              "$ref": "#/components/schemas/SensorReadingOut"
            },
            "type": "array",
            "title": "Readings"
          },
          "summary": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "field_id",
          "sensor_id",
          "readings_created",
          "start_at",
          "end_at",
          "readings",
          "summary"
        ],
        "title": "SensorSimulationResponse"
      },
      "SensorTrend": {
        "properties": {
          "field_id": {
            "type": "integer",
            "title": "Field Id"
          },
          "sensor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor Id"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/SensorTrendPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "statistics": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Statistics"
          },
          "quality_issue_count": {
            "type": "integer",
            "title": "Quality Issue Count"
          },
          "latest_reading": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SensorReadingOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "field_id",
          "sensor_id",
          "points",
          "statistics",
          "quality_issue_count"
        ],
        "title": "SensorTrend"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}